Restructured people to be easier
Assuming each person has their own directory doesn't really make a lot of
sense to me, unless you have a good reason. If you do, we can move them
back in. But this way, we just loop through the individual files and
output their HTML directly, instead of ending up with a bunch of sub-files
that are all (confusingly) named the same thing.