Adds EJS plugin (Underscore templating)
@chrisburkedev

Don't merge from this, but you might find it useful as an example.

It demonstrates using the EJS templating, and it creates an `include()` function that can be
used to pull in sub-templates with a new context object that inherits prototypically from
the old context object (meaning that you can access both the new and old values). There
might be bugs in there.